requirements.txt
PyHamcrest==1.8.0
lxml==3.3.3
